+++ src/Log/Writer/ContextException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Facile\SentryModule\Log\Writer;
+
+class ContextException extends \RuntimeException
+{
+}